2025-12-05_super_taller_cv/python/Subsistema_1-Deteccion_y_segmentacion_inteligente/requirements.txt
# Core Detection and Segmentation
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ultralytics>=8.0.0
mediapipe>=0.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git

# Computer Vision Extras
pillow>=10.0.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.28.0
streamlit-webrtc>=0.45.0

# Web & Communication
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
requests>=2.31.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional for GPU
# onnxruntime-gpu>=1.16.0

# Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0

# Export formats
openpyxl>=3.1.0